requirements.txt
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.5.0
plotly==5.13.1
streamlit==1.20.0
